+++ src/main/java/com/project/vehiclerental/advice/exception/FileDownloadException.java
@@ -1,4 +1,4 @@
-package com.project.vehiclerental.exception;
+package com.project.vehiclerental.advice.exception;
 
 public class FileDownloadException extends SpringBootFileUploadException {
     public FileDownloadException(String message) {
