+++ src/main/java/com/project/vehiclerental/advice/exception/FileEmptyException.java
@@ -1,4 +1,4 @@
-package com.project.vehiclerental.exception;
+package com.project.vehiclerental.advice.exception;
 
 public class FileEmptyException extends SpringBootFileUploadException {
     public FileEmptyException(String message) {
