+++ src/main/java/com/project/vehiclerental/advice/exception/FileUploadException.java
@@ -1,4 +1,4 @@
-package com.project.vehiclerental.exception;
+package com.project.vehiclerental.advice.exception;
 
 public class FileUploadException extends SpringBootFileUploadException {
     public FileUploadException(String message) {
