+++ src/main/java/com/project/vehiclerental/service/FileService.java
@@ -1,7 +1,7 @@
 package com.project.vehiclerental.service;
 
-import com.project.vehiclerental.exception.FileDownloadException;
-import com.project.vehiclerental.exception.FileUploadException;
+import com.project.vehiclerental.advice.exception.FileDownloadException;
+import com.project.vehiclerental.advice.exception.FileUploadException;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
